Add regression test for issue #157
This test ensures that when `interrupt()` is called, the Event
Handler
thread is not calling `dataFileWriter.close()`.

In order to have a reproducable test and avoid flakyness, we m
ock the
following methods to ensure the order and timming of the race
condition:

- `dataFileWriter.close()` is locked on a latch until an `interrupt()`
is sent, if that interrupt happens inside this call, it will cause an
assertion fail;
- `interrupt()` unlocks that latch after executing the real
implementation;
- `eventHandlerThread` requires `dataFileWriter` to be set with the
mock, as well as `inProgressHistFile` to be set with a real path, to
pass some checks during `stop()`.